add tk function for easier session killing
tk function is not bash portable right now - it uses the special
read syntax for zsh. i've been meaning to break out common code
into a different profile, and this is a good reason to do that -
but for now, this'll work.

see this link for details on implementation differences:
https://superuser.com/questions/555874/zsh-read-command-fails-within-bash-function-read1-p-no-coprocess