break some env vars out into more appropriate files
this is almost certainly not done, did it kinda slapdash, but this
state at least gets alchemist working again, so it'll do for now.
probably the actual solution to this will involve separating common env
settings and common profile shit like functions and aliases into
different files, and sourcing those from the appropriate spots. i've
made a bit of a mess of it for now but whatever.