tmux splits open in cwd from previous pane
this is a change to the default behavior where all splits use the cwd
where the tmux command to start the session was originally run. it seems
like what i usually want when making a new window or split is to be in
the same context i'm splitting from, so, now that's the case.